+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Add email settings link
 * Begin selecting payments for processing
 * Enable activation and version checking
+* Link to order details during processing
+* Move processor to Downloads menu section
 * Remove donate links
 * Removed unused EDD_Retroactive_Licensing_Settings
 * Update description